(engine) split up the engine internals to own files/modules
